--- kuka_lbr_iiwa7_support/urdf/lbriiwa7.xacro
@@ -1,5 +1,0 @@
-<?xml version="1.0"?>
-<robot name="lbriiwa7" xmlns:xacro="http://wiki.ros.org/xacro">
-  <xacro:include filename="$(find kuka_lbr_iiwa7_support)/urdf/lbriiwa7_macro.xacro"/>
-  <xacro:kuka_lbr_iiwa7 io_access="false"/>
-</robot>
