+++ kuka_lbr_iiwa_support/meshes/lbr_iiwa14_r820/visual/iiwa_r820_link_0_hull_v0.1_iiwa_r820_link_0_hull.stl
