+++ kuka_lbr_iiwa_support/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>kuka_lbr_iiwa7_support</name>
+  <name>kuka_lbr_iiwa_support</name>
   <version>0.0.0</version>
-  <description>ROS2 support package for the KUKA LBR iiwa 7</description>
+  <description>ROS2 support package for the KUKA LBR iiwa</description>
   <maintainer email="svastits.aron@gmail.com">Áron Svastits</maintainer>
   <license>Apache 2.0</license>
